+++ lib/colored_enumeration/extensions/enumeration_format_patch.rb
@@ -2,7 +2,7 @@
 
 # This file is part of the Plugin Redmine Colored Enumeration.
 #
-# Copyright (C) 2021 Liane Hampe <liaham@xmera.de>, xmera.
+# Copyright (C) 2021 - 2022 Liane Hampe <liaham@xmera.de>, xmera.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -23,11 +23,23 @@ module ColoredEnumeration
     module EnumerationFormatPatch
       def self.included(base)
         base.include(InstanceMethods)
+        base.multiple_supported = false
       end
 
       module InstanceMethods
+        ##
+        # Strips the color of a custom field value.
+        #
+        # @note Given a multi value field the first color will be taken.
+        #
+        # @return [String] A hex color string.
+        #
         def cast_color(custom_field, value, customized = nil)
-          value_object(custom_field, value, customized)&.color || ''
+          casted_value = value_object(custom_field, value, customized)
+          return casted_value&.color || '' unless casted_value.is_a? Array
+
+          colors = casted_value&.map(&:color)
+          colors.first || ''
         end
 
         def value_object(custom_field, value, customized)