+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.2 - 2022-04-28
+
+### Changed
+
+* copyright year
+
+### Fixed
+
+* color casting for multiple value fields
+* support for colored enumerations in issue list
+
 ## 0.1.1 - 2021-12-03
 
 ### Fixed